Remove the tracker from the client config.
Since the server is already in control of the torrent files it can create
torrents which point to any tracker it chooses. The plan is to have a script on
the server which accepts a bundle and generates a torrent for it, presumably
with the default tracker set to itself.

Also change the name of the attribute which gives the torrent file, to make it
clear that it is the base of a URL which links to the torrent file.